2007-04-23 18:25 UTC+0200 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * harbour/include/hbcomp.h   * harbour/include/hbcompdf.h   * harbour/include/hbexpra.c   * harbour/include/hbexprb.c   * harbour/include/hbexprop.h   * harbour/source/common/expropt1.c   * harbour/source/compiler/genc.c   * harbour/source/compiler/harbour.y   * harbour/source/compiler/hbmain.c   * harbour/source/compiler/hbopt.c     ! reverted fix for unnecessary -W2 warning:          Function '...' does not end with RETURN statement       in code like:             func f(v)             if v                return "TRUE"             else                return "FALSE"             endif       Sorry but I had to make it because this fix was wrong and causes       that code like:             func f(v)             if v                ? v             else                return "FALSE"             endif       was compiled without any warnings but wrong PCODE was generated.       In some spare time I'll try to implement valid RETURN detection,       now simply add RETURN NIL at the end of such functions - it will       be removed by dead code eliminator.     ! fixed memory leak when more then one .prg file was given as compiler       parameter     + cleaned some code for future modifications
  * harbour/source/macro/macro.y
    + added small hack for two BCC/OW warnings
  * harbour/source/compiler/harbour.yyc
  * harbour/source/compiler/harbour.yyh
  * harbour/source/macro/macro.yyc
  * harbour/source/macro/macro.yyh
    * regenerated with modified Bison version which should give code
      without BCC warnings - I cannot promise I'll keep it in the
      future but I will try.

  * harbour/source/rdd/dbcmd.c
    ! use default RDD instead of current one in COPY TO and APPEND FROM

  * harbour/source/vm/classes.c
    * added support for (@func()):eval(...)

+++ harbour/include/hbexprop.h
@@ -89,7 +89,7 @@ extern HB_EXPR_PTR hb_compExprNewDate( HB_LONG, HB_COMP_DECL );
 extern HB_EXPR_PTR hb_compExprNewString( char *, ULONG, BOOL, HB_COMP_DECL );
 extern HB_EXPR_PTR hb_compExprNewLogical( int, HB_COMP_DECL );
 extern HB_EXPR_PTR hb_compExprNewSelf( HB_COMP_DECL );
-extern HB_EXPR_PTR hb_compExprNewCodeBlock( char *, int, int, HB_COMP_DECL );
+extern HB_EXPR_PTR hb_compExprNewCodeBlock( char *, ULONG, int, HB_COMP_DECL );
 extern HB_EXPR_PTR hb_compExprNewVar( char *, HB_COMP_DECL );
 extern HB_EXPR_PTR hb_compExprNewAliasVar( HB_EXPR_PTR, HB_EXPR_PTR, HB_COMP_DECL );
 extern HB_EXPR_PTR hb_compExprNewAliasExpr( HB_EXPR_PTR, HB_EXPR_PTR, HB_COMP_DECL );
